modify server return type of (anonymousRequest,endorsement,printerIssue) to be of same type, finish anonymousRequest CRUD

+++ src/resources/actions/general/endorsement/endorsement.controller.ts
@@ -1,6 +1,7 @@
 import expressAsyncHandler from 'express-async-handler';
 import {
   createNewEndorsementService,
+  deleteAllEndorsementsService,
   deleteEndorsementService,
   getAllEndorsementsService,
   getAnEndorsementService,
@@ -10,6 +11,7 @@ import {
 import type { Response } from 'express';
 import type {
   CreateNewEndorsementRequest,
+  DeleteAllEndorsementsRequest,
   DeleteEndorsementRequest,
   GetAllEndorsementsRequest,
   GetAnEndorsementRequest,
@@ -30,8 +32,6 @@ const createNewEndorsementHandler = expressAsyncHandler(
       summaryOfEndorsement,
     } = request.body;
 
-    console.log('request.body', request.body);
-
     const newEndorsementObject = {
       userId,
       section,
@@ -46,10 +46,12 @@ const createNewEndorsementHandler = expressAsyncHandler(
     if (newEndorsement) {
       response.status(201).json({
         message: 'Endorsement created successfully',
-        endorsement: newEndorsement,
+        endorsementData: [newEndorsement],
       });
     } else {
-      response.status(400).json({ message: 'Endorsement could not be created', endorsement: {} });
+      response
+        .status(400)
+        .json({ message: 'Endorsement could not be created', endorsementData: [] });
     }
   }
 );
@@ -59,14 +61,28 @@ const createNewEndorsementHandler = expressAsyncHandler(
 // @access Private
 const getAllEndorsementsHandler = expressAsyncHandler(
   async (request: GetAllEndorsementsRequest, response: Response) => {
+    // only managers/admins can view all endorsements
+    const {
+      userInfo: { roles },
+    } = request.body;
+    if (!roles.includes('manager') && !roles.includes('admin')) {
+      response.status(403).json({
+        message: 'Only managers and admins can view all endorsements',
+        endorsementData: [],
+      });
+      return;
+    }
+
     const endorsements = await getAllEndorsementsService();
 
     if (endorsements.length === 0) {
-      response.status(400).json({ message: 'Endorsements could not be fetched', endorsements: [] });
+      response
+        .status(400)
+        .json({ message: 'Endorsements could not be fetched', endorsementData: [] });
     } else {
       response.status(200).json({
         message: 'Endorsements fetched successfully',
-        endorsements,
+        endorsementData: [endorsements],
       });
     }
   }
@@ -77,16 +93,30 @@ const getAllEndorsementsHandler = expressAsyncHandler(
 // @access Private
 const getAnEndorsementHandler = expressAsyncHandler(
   async (request: GetAnEndorsementRequest, response: Response) => {
+    // only managers/admins can view an endorsement by its id
+    const {
+      userInfo: { roles },
+    } = request.body;
+    if (!roles.includes('manager') && !roles.includes('admin')) {
+      response.status(403).json({
+        message: 'Only managers and admins can view an endorsement not belonging to them',
+        endorsementData: [],
+      });
+      return;
+    }
+
     const { endorsementId } = request.params;
 
     const endorsement = await getAnEndorsementService(endorsementId);
     if (endorsement) {
       response.status(200).json({
         message: 'Endorsement fetched successfully',
-        endorsement,
+        endorsementData: [endorsement],
       });
     } else {
-      response.status(400).json({ message: 'Endorsement could not be fetched', endorsement: [] });
+      response
+        .status(400)
+        .json({ message: 'Endorsement could not be fetched', endorsementData: [] });
     }
   }
 );
@@ -96,17 +126,18 @@ const getAnEndorsementHandler = expressAsyncHandler(
 // @access Private
 const getEndorsementsByUserHandler = expressAsyncHandler(
   async (request: GetEndorsementsFromUserRequest, response: Response) => {
+    // anyone can view their own endorsements
     const {
       userInfo: { userId },
     } = request.body;
 
     const endorsements = await getEndorsementsByUserService(userId);
     if (endorsements.length === 0) {
-      response.status(400).json({ message: 'No endorsements found', endorsements: [] });
+      response.status(400).json({ message: 'No endorsements found', endorsementData: [] });
     } else {
       response.status(200).json({
         message: 'Endorsements fetched successfully',
-        endorsements,
+        endorsementData: [endorsements],
       });
     }
   }
@@ -117,16 +148,61 @@ const getEndorsementsByUserHandler = expressAsyncHandler(
 // @access Private
 const deleteEndorsementHandler = expressAsyncHandler(
   async (request: DeleteEndorsementRequest, response: Response) => {
+    // only managers/admins can delete an endorsement by its id
+    const {
+      userInfo: { roles },
+    } = request.body;
+    if (!roles.includes('manager') && !roles.includes('admin')) {
+      response.status(403).json({
+        message: 'Only managers and admins can delete an endorsement',
+        endorsementData: [],
+      });
+      return;
+    }
+
     const { endorsementId } = request.params;
 
     const deletedEndorsement = await deleteEndorsementService(endorsementId);
     if (deletedEndorsement) {
       response.status(200).json({
         message: 'Endorsement deleted successfully',
-        endorsement: deletedEndorsement,
+        endorsementData: [deletedEndorsement],
       });
     } else {
-      response.status(400).json({ message: 'Endorsement could not be deleted', endorsement: {} });
+      response
+        .status(400)
+        .json({ message: 'Endorsement could not be deleted', endorsementData: [] });
+    }
+  }
+);
+
+// @desc   Delete all endorsements
+// @route  DELETE /endorsements
+// @access Private
+const deleteAllEndorsementsHandler = expressAsyncHandler(
+  async (request: DeleteAllEndorsementsRequest, response: Response) => {
+    // only managers/admins can delete all endorsements
+    const {
+      userInfo: { roles },
+    } = request.body;
+    if (!roles.includes('manager') && !roles.includes('admin')) {
+      response.status(403).json({
+        message: 'Only managers and admins can delete all endorsements',
+        endorsementData: [],
+      });
+      return;
+    }
+
+    const deletedEndorsements = await deleteAllEndorsementsService();
+    if (deletedEndorsements.acknowledged) {
+      response.status(200).json({
+        message: 'Endorsements deleted successfully',
+        endorsementData: [],
+      });
+    } else {
+      response
+        .status(400)
+        .json({ message: 'Endorsements could not be deleted', endorsementData: [] });
     }
   }
 );
@@ -135,6 +211,7 @@ export {
   createNewEndorsementHandler,
   getAnEndorsementHandler,
   deleteEndorsementHandler,
+  deleteAllEndorsementsHandler,
   getAllEndorsementsHandler,
   getEndorsementsByUserHandler,
 };
